requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
faker==25.2.0
openpyxl==3.1.2
python-dotenv==1.0.1